Make use of auto-close on strnlen()
Like the tutorial says, when working with loop contracts (instead of
loop invariants) it's better to put the condition inside the loop
to be able to use ghost open/close commands. In addition now we use
auto close at the exit points of the loop. This removes the potential
arithmetic overflow error verifast complained about.